Revise packaging and testing.

+++ admin-tools/make-dist-newest.sh
@@ -1,5 +1,14 @@
 #!/bin/bash
-PACKAGE=xdis
+# The name Python's import uses.
+# It is reflected in the directory structure.
+PACKAGE_MODULE="xdis"
+
+# The name that PyPi sees this as.
+# It is set in setup.py's name.
+PACKAGE_NAME="xdis"
+
+# Both the name an module name agree
+PACKAGE=$PACKAGE_NAME
 
 # FIXME put some of the below in a common routine
 function finish {
@@ -21,10 +30,39 @@ fi
 
 cd ..
 source $PACKAGE/version.py
-echo $__version__
-pyenv local 3.13
+if [[ ! $__version__ ]] ; then
+    echo "Something is wrong: __version__ should have been set."
+    exit 1
+fi
 
-rm -fr build
-pip wheel --wheel-dir=dist .
+for pyversion in $PYVERSIONS; do
+    if [[ ${pyversion:0:4} == "pypy" ]] ; then
+	echo "$pyversion - PyPy does not get special packaging"
+	continue
+    fi
+    echo "*** Packaging ${PACKAGE_NAME} for version ${__version__} on Python ${pyversion} ***"
+    if ! pyenv local $pyversion ; then
+	exit $?
+    fi
+    # pip bdist_egg create too-general wheels. So
+    # we narrow that by moving the generated wheel.
+
+    # Pick out first two number of version, e.g. 3.5.1 -> 35
+    first_two=$(echo $pyversion | cut -d'.' -f 1-2 | sed -e 's/\.//')
+    rm -fr build
+    pip wheel --wheel-dir=dist .
+    mv -v dist/${PACKAGE_MODULE}-$__version__-{py3,py$first_two}-none-any.whl
+done
+
 python -m build --sdist
+tarball=dist/${PACKAGE_NAME}-${__version__}.tar.gz
+if [[ -f $tarball ]]; then
+    twine check $tarball
+fi
+
+if [[ ! -d dist/${__version__} ]] ; then
+    mkdir -v dist/${__version__}
+fi
+
+twine check dist/${PACKAGE}-${__version__}-py3*.whl
 finish
